Feat: Remover página de login e componente de autenticação, adicionar nova estrutura de login
