--- vozes-na-tech/app/Page/LoginPage.tsx
@@ -1,10 +1,0 @@
-"use client";
-import React from "react";
-import Login from "@/app/components/login";
-
-
-function LoginPage() {
-    return <Login />;
-}
-
-export default LoginPage;
