+++ vozes-na-tech/app/home/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+import React from "react";
+import Inicial from "@/app/page"
+
+
+export default function Home() {
+return <Inicial />
+
+}
